+++ packages/dapp/src/components/Web3AuthWrapper.tsx
@@ -6,8 +6,11 @@ import AuthWeb3Login from "./Auth/AuthWeb3Login";
 import SetUsername from "./Auth/SetUsername";
 import SetEmail from "./Auth/SetEmail";
 import { hideWeb3AuthModal, showWeb3LoginModal, showWeb3SignupModal } from "../redux/actionCreators/ui";
+import { setNetwork, setNetworkName } from "../redux/actionCreators/network";
 import { ICivilContext, CivilContext } from "@joincivil/components";
+import { setNetworkValue, setDefaultNetworkValue } from "@joincivil/utils";
 import SetAvatar from "./Auth/SetAvatar";
+import config from "../helpers/config";
 
 export const Web3AuthWrapper: React.FunctionComponent = () => {
   // context
@@ -25,6 +28,45 @@ export const Web3AuthWrapper: React.FunctionComponent = () => {
       state.networkDependent.user.account &&
       state.networkDependent.user.account.account,
   );
+
+  React.useEffect(() => {
+    setDefaultNetworkValue(parseInt(config.DEFAULT_ETHEREUM_NETWORK!, 10));
+    const civil = civilContext.civil!;
+    const networkSub = civil.networkStream.subscribe(onNetworkUpdated);
+    const networkNameSub = civil.networkNameStream.subscribe(onNetworkNameUpdated);
+    const accountSub = civil.accountStream.subscribe(onAccountUpdated);
+
+    return function cleanup(): void {
+      networkSub.unsubscribe();
+      networkNameSub.unsubscribe();
+      accountSub.unsubscribe();
+    };
+  }, [civilContext]);
+
+  async function onNetworkUpdated(network: number): Promise<void> {
+    dispatch!(setNetwork(network.toString()));
+    setNetworkValue(network);
+  }
+
+  async function onNetworkNameUpdated(networkName: string): Promise<void> {
+    dispatch!(setNetworkName(networkName));
+  }
+
+  const onAccountUpdated = async (account: string | undefined): Promise<void> => {
+    const currentUser = civilContext.currentUser;
+    if (
+      account &&
+      currentUser &&
+      currentUser.ethAddress &&
+      account.toLowerCase() !== currentUser.ethAddress.toLowerCase()
+    ) {
+      console.warn("web3 account does not match that of the logged in user, logging out", {
+        web3Account: account,
+        currentUserAccount: currentUser.ethAddress,
+      });
+      civilContext.auth.logout();
+    }
+  };
 
   const showWeb3Signup = showWeb3AuthModal && web3AuthType === "signup";
   const showWeb3Login = showWeb3AuthModal && web3AuthType === "login";
